Fix width of _stp_filename_lookup_5's offset to .debug_line_str
The read_pointer( ... DW_EH_PE_data4 ... ) gives a 4 byte value.
Elfutil's readelf.c does this with read_4ubyte_unaligned_inc().
Adjust the storage width for such offset to prevent overflow.
Problem demonstrated with context.exp / symfileline.tcl.

+++ runtime/common_probe_context.h
@@ -212,6 +212,6 @@ unsigned long kern_bt_entries[MAXBACKTRACE];
 struct {
   struct { uint16_t desc; uint16_t form; } dir_enc[STP_MAX_DW_SOURCES];
   struct { uint16_t desc; uint16_t form; } file_enc[STP_MAX_DW_SOURCES];
-  struct { uint8_t offset; char *name; } src_dir[STP_MAX_DW_SOURCES];
-  struct { uint8_t offset; uint8_t dirindex; char *name; } src_file[STP_MAX_DW_SOURCES];
+  struct { uint32_t offset; char *name; } src_dir[STP_MAX_DW_SOURCES];
+  struct { uint32_t offset; uint8_t dirindex; char *name; } src_file[STP_MAX_DW_SOURCES];
 } dw_data;
